[#211] Update top-level namespace to Twitter::TwitterText::
The popular Twitter gem (https://github.com/sferik/twitter) uses the
Twitter::Configuration namespace and users are reporting that our new
Twitter::Configuration class is conflicting with that namespace.

This patch changes our top-level namespace from Twitter to
Twitter::TwitterText:: to resolve this conflict and prevent further
conflicts in the future. Because this is a breaking change for clients
written against older versions of the library, the version has been
bumped to 2.1.

There's also a new CHANGELOG.md file that details important changes in
this revision and will be updated going forward.